==> survey/analysis/8.a_few_extra_stats.ipynb <==
# Count number of responses for various topics
The tagging of comments makes it such that we must add together a few different sub-counts to get the total

import numpy as np
import pandas as pd
from IPython.display import display, Markdown, Latex
df = pd.read_csv('responses_complete.csv', index_col=None, parse_dates=['date'])

#don't truncate results
pd.set_option('display.max_rows', None) # display all rows
pd.set_option('display.max_colwidth', -1) # display full text of columns


## Number of comments about 129/Maple

filtered = (df['problem_tags_maple'] == 1) | (df['problem_tags_129'] == 1) | (df['problem_tags_van wyck'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about S Riverside

filtered = (df['problem_tags_s riverside'] == 1) | (df['problem_tags_shoprite'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about Cleveland

filtered = (df['problem_tags_cleveland'] == 1) | (df['problem_tags_five corners'] == 1) | (df['problem_tags_cet'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about Benedict Boulevard


filtered = (df['problem_tags_benedict'] == 1) | (df['problem_tags_gottwald circle'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about the Dummy Light

filtered = df['problem_tags_benedict'] == 1) | (df['problem_tags_gottwald circle'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about sidewals

filtered = (df['problem_tags_sidewalk condition'] == 1) | (df['problem_tags_missing sidewalks'] == 1) | (df['problem_tags_bushes on sidewalk'] == 1) | (df['suggestion_tags_more sidewalks'] == 1) | (df['suggestion_tags_maintain sidewalks'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about crosswalks

filtered = (df['problem_tags_missing crosswalks'] == 1) | (df['suggestion_tags_maintain crosswalks'] == 1) | (df['suggestion_tags_more crosswalks'] == 1)
df2 = df[filtered]
display(df2.shape[0])

## Number of comments about driver behavior

filtered = (df['problem_tags_aggressive driving'] == 1) | (df['problem_tags_speeding'] == 1) | (df['problem_tags_driver awareness'] == 1) | (df['suggestion_tags_enforce speed'] == 1) | (df['suggestion_tags_reduce speed'] == 1) | (df['suggestion_tags_speed bumps'] == 1) | (df['suggestion_tags_speed cameras'] == 1) | (df['suggestion_tags_traffic calming'] == 1)    
df2 = df[filtered]
display(df2.shape[0])

## Insights into school bus ridership

### Number respondents who say they don't have school bus service available

num_no_bus = df['child_drive_reason_Lack of available busing where we live'].sum()
num_no_bus

### Frequency with which students take the bus

df['child_bus_freq'].value_counts()

df['child_bus_freq'].value_counts(normalize=True).mul(100).round(1).astype(str) + '%'

### Number of respondents who told us about their school-aged children

total_respondents_with_kids = df['child_bus_freq'].value_counts().sum()
total_respondents_with_kids

### Estimate of how many respondents don't have bus service where they live
Based on our understanding that the school district has 1600 students, including 1,048 with bus service, we estimate that around 35% do not have bus service available where they live.

estimate_no_bus_service = round(total_respondents_with_kids * 0.35)
estimate_no_bus_service

estimate_no_bus_service_as_percent = round((estimate_no_bus_service / total_respondents_with_kids) * 100)
str(estimate_no_bus_service_as_percent) + "%"

### Adjustment to account for estimated number of students who don't have bus service available

df2 = df['child_bus_freq'].value_counts()
# subtract estimated number of students without bus service from the 'rarely or never' response count
df2.iloc[1] = df2.iloc[1] - estimate_no_bus_service 
df2

Total number of respondents in our revised estimate

total_respondents_with_kids = df2.sum()
total_respondents_with_kids

df2.apply(lambda x: str(round(x/df2.sum() * 100)) + '%' )
